Cambio en la estructura, programa dentro de un directoio

+++ metro-laser/env.h
@@ -6,4 +6,4 @@ const byte DS3231_I2C_ADDRESS = 0x68;
 // SD
 
 // ARDUINO
-const int BOTON = 2; //Pin de la botonera usada para almacenar los datos
+const byte BOTON = 2; //Pin de la botonera usada para almacenar los datos
